refactor: Centralize and consolidate constants
Moved all constants to the `constants` directory, organized by domain.
Updated all references to the moved constants.
Deleted the now-empty `lib/spotify.ts` file.
Moved `refreshSpotifyToken` and `getSpotifyBasicAuth` to `lib/spotify-auth.ts` to fix a circular dependency in the tests.
Fixed all broken tests.
