+++ constants/ui.ts
@@ -1,0 +1,2 @@
+// constants/ui.ts
+export const SIDE_COLUMN_WIDTH = '40px'
