+++ context/webSocketReducer.ts
@@ -5,6 +5,7 @@ import {
   ActiveAlert,
 } from '../types/websocket'
 import { HrmStreamData as ServerHrmData } from '../types/core'
+import { INITIAL_STATE } from '@/constants/webSocket'
 
 // Client-side extension of HrmData to include connection status
 export interface HrmData extends ServerHrmData {
@@ -17,34 +18,6 @@ export interface WebSocketState {
   spotifyData: SpotifyData
   activeAlerts: ActiveAlert[]
   spotifyServiceInitialized?: boolean
-}
-
-export const INITIAL_STATE: WebSocketState = {
-  hrmData: [],
-  timerData: {
-    isRunning: false,
-    currentPhase: 'IDLE',
-    timeRemaining: 0,
-    timeElapsed: 0,
-    caloriesBurned: 0,
-    mode: 'TABATA',
-    workDuration: 30,
-    restDuration: 10,
-    soundEventId: 0,
-  },
-  spotifyData: {
-    trackId: null,
-    trackName: 'Awaiting Login...',
-    artist: '',
-    albumName: '',
-    albumArtUrl: '',
-    isPlaying: false,
-    devices: [],
-    volume: 70,
-    isMuted: false,
-  },
-  activeAlerts: [],
-  spotifyServiceInitialized: false,
 }
 
 export const reducer = (
